Made LegacyCommandRunner simply mimic CommandRunner interface
Created LegacyShellDispatcher again as it is the only way to inject the
IO mock into the shell, but we cannot add a non-static run() method
because ShellDispatcher has a static::run
